remove "build at head"
We're finding that the build system is fuzzy with respect to the point at
which it builds amongst a set of possible tags and commits and HEAD on a
branch.  It is better to be explicit for now, given the types of errors
we're seeing.

This patch removes the mostly implicit "build at head" behavior and results
in a --buildversion being required.  Without a --buildversion the system
will propose one to you.